Add recognition pipeline architecture documentation
- Introduced a new markdown file detailing the recognition pipeline architecture, including threading rules and shared state conventions for the EndFieldFightHelper.
- Documented key components such as Capture and Detection threads, shared state models, and their interactions.
- Provided guidelines for adding new pipeline stages and managing shared state effectively.
- Enhanced understanding of the recognition pipeline's lifecycle and API for developers.

+++ Helpers/Win32Helper.cs
@@ -199,11 +199,11 @@ public static class Win32Helper
         var styles = GetWindowLongPtr(hWnd, GWL_EXSTYLE).ToInt64();
         if (enabled)
         {
-            styles |= exStyle;
+            styles |= (uint)exStyle;
         }
         else
         {
-            styles &= ~exStyle;
+            styles &= ~(long)(uint)exStyle;
         }
 
         SetWindowLongPtr(hWnd, GWL_EXSTYLE, new IntPtr(styles));
